pyproject.toml
[tool.poetry]
name = "python-socket-io-implement"
version = "0.1.0"
description = ""
authors = ["Prem Kumar G <premkumarmvks@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-socketio = "^5.11.1"
gunicorn = "^21.2.0"
eventlet = "^0.35.1"
uvicorn = "^0.27.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
